requirements.txt
garminconnect
garth
google-api-python-client
google-auth
google-auth-oauthlib
openpyxl
python-dotenv
